# feat: Speed up adding native platform